feat: implements owner and repositories input for token generation (#277)
This PR adds the `owner` and `repositories` parameters inherited from
the official GitHub action.

By default, the generated tokens have a scope limited to the current
repository. To extend this scope to other repositories, you can list
them in the `repositories` field. If you want the scope to apply to all
repositories of the owner, you should specify the value `!null`.

I have tested this in the following workflows:
- [Run
11220366350](https://github.com/camunda/team-infrastructure-experience/actions/runs/11220366350/job/31188374915)
- [Run
11220233216](https://github.com/camunda/team-infrastructure-experience/actions/runs/11220233216/job/31187940158)

---------

Co-authored-by: Maxim Danilov <maxdanilov@users.noreply.github.com>