fix: retry-worfklow-run report back to workflow file
Before it was just taking the workflowName, which can be the correct fileName if a user doesn't overwrite the name. In that case it will default to the file name.
The likelihood of users doing that is quite slim.
Therefore, we need to get the correct file name somehow differently as the run itself doesn't contain that, we're parsing it via the api.