chore: allow retry action and workflow to rerun whole workflow (#463)
Hey Infra 👋,

I have a small suggestion: Allow the retry action/workflow to retrigger
the whole workflow, not just the failed jobs.

We have a use case where it would be beneficial for us to rerun the
whole workflow instead of just the failed jobs. Since we already have
the functionality available, I thought to simply extend it instead of
recreating it on our side.

For existing users nothing changes as the default is to retry the failed
jobs.

Related [gh run rerun docs](https://cli.github.com/manual/gh_run_rerun).

> Rerun an entire run, only failed jobs, or a specific job from a run.

Let me know if names should be changed or implemented differently.
We could also approach it from a different side to say
`rerun-failed-jobs` or similar and make it the default to be true and
have something like `EXTRA_ARGS` that's passed in that's either empty or
`failed` and maybe extensible as extra input to allow passing in
anything from outside.